switch to using hosted image of a cat-dev
in the past i encoded a gigantic jpeg as Base64 just because I really
wanted _everyone_ to be able to edit the image and preview the image.
unfortunately, most sites (github included) choose to not allow base64
encoded images (or maybe not as large of base64 encoded images).

this means while anyone could submit a pr for it, it just wouldn't
render on most sites. making it useless. now it's uploaded to one of my
loose backblaze buckets, which means not everyone can techincally own
it, but folks can submit a pr, i upload the image, and then the pr
works. OR we can choose a more community owned url mirror later on too.
it doesn't need to be my bucket.

this will at least let the image render on pretty much every site.