write as needed + resumption
fixes #16

this writes the memory dump file which takes forever (4Gbs at 512
bytes per request, which can sometimes fail and need to be retried)
as the requests come in rather than actually waiting for it to all be
buffered in memory first. this also lowers the total amount of memory
required to do so.

also allow resuming a stopped/failed memory dump.