introduce mion dump-eeprom & dump-memory (#15)
* introduce mion dump-eeprom & dump-memory

introduce two new subcommands we found while poking around in telnet
memory dumps (telnet  has the same username/password as the http
interface). these cgi pages allow us to figure out the memory, and
eeprom of the MION itself, which should be helpful for certain tasks.

* retry dumping memory

* fix clap breakage

* allow retry on body read too

* write as needed + resumption

fixes #16

this writes the memory dump file which takes forever (4Gbs at 512
bytes per request, which can sometimes fail and need to be retried)
as the requests come in rather than actually waiting for it to all be
buffered in memory first. this also lowers the total amount of memory
required to do so.

also allow resuming a stopped/failed memory dump.

* mention how long a mion memory dump takes

* concurrently fetch pages