fix rename dir regression for windows
Windows will often complain about removing the symlink to a directory
even if the user created the symlink, and has permissions to both
directories.

as a result, we now won't attempt to delete the symlink, and let the
OS clean it up itself in the final remove directory that is being
called.